feat: refactor endpoints with 'cognito' to new URIs
OCD-4660
