+++ chpl/chpl-api/src/main/java/gov/healthit/chpl/web/controller/AuthenticationController.java
@@ -1,5 +1,9 @@
 package gov.healthit.chpl.web.controller;
 
+import java.util.UUID;
+
+import org.apache.commons.lang3.NotImplementedException;
+import org.ff4j.FF4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.security.core.Authentication;
@@ -14,12 +18,19 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.nulabinc.zxcvbn.Strength;
 
+import gov.healthit.chpl.FeatureList;
 import gov.healthit.chpl.auth.ChplAccountEmailNotConfirmedException;
 import gov.healthit.chpl.auth.ChplAccountStatusException;
 import gov.healthit.chpl.auth.authentication.JWTUserConverterFacade;
 import gov.healthit.chpl.auth.user.AuthenticationSystem;
 import gov.healthit.chpl.auth.user.JWTAuthenticatedUser;
+import gov.healthit.chpl.domain.CognitoRefreshTokenRequest;
 import gov.healthit.chpl.domain.auth.AuthenticationResponse;
+import gov.healthit.chpl.domain.auth.CognitoForgotPasswordRequest;
+import gov.healthit.chpl.domain.auth.CognitoLogoutRequest;
+import gov.healthit.chpl.domain.auth.CognitoNewPasswordRequiredRequest;
+import gov.healthit.chpl.domain.auth.CognitoSetForgottenPasswordRequest;
+import gov.healthit.chpl.domain.auth.CognitoUpdatePasswordRequest;
 import gov.healthit.chpl.domain.auth.LoginCredentials;
 import gov.healthit.chpl.domain.auth.ResetPasswordRequest;
 import gov.healthit.chpl.domain.auth.UpdatePasswordRequest;
@@ -33,9 +44,15 @@ import gov.healthit.chpl.exception.JWTValidationException;
 import gov.healthit.chpl.exception.MultipleUserAccountsException;
 import gov.healthit.chpl.exception.UserManagementException;
 import gov.healthit.chpl.exception.UserRetrievalException;
+import gov.healthit.chpl.exception.ValidationException;
 import gov.healthit.chpl.manager.auth.AuthenticationManager;
 import gov.healthit.chpl.manager.auth.UserManager;
 import gov.healthit.chpl.service.UserAccountUpdateEmailer;
+import gov.healthit.chpl.user.cognito.authentication.CognitoAuthenticationChallengeException;
+import gov.healthit.chpl.user.cognito.authentication.CognitoAuthenticationManager;
+import gov.healthit.chpl.user.cognito.authentication.CognitoAuthenticationResponse;
+import gov.healthit.chpl.user.cognito.authentication.CognitoPasswordResetRequiredException;
+import gov.healthit.chpl.user.cognito.password.CognitoPasswordManager;
 import gov.healthit.chpl.util.AuthUtil;
 import gov.healthit.chpl.util.ErrorMessageUtil;
 import gov.healthit.chpl.util.SwaggerSecurityRequirement;
@@ -59,21 +76,142 @@ public class AuthenticationController {
     private JWTUserConverterFacade userConverterFacade;
     private ErrorMessageUtil msgUtil;
 
+    private CognitoAuthenticationManager cognitoAuthenticationManager;
+    private CognitoPasswordManager cognitoPasswordManager;
+    private FF4j ff4j;
+
+
     @Autowired
-    public AuthenticationController(AuthenticationManager authenticationManager,
-            UserAccountUpdateEmailer userAccountUpdateEmailer,
-            BCryptPasswordEncoder bCryptPasswordEncoder,
-            UserManager userManager, JWTUserConverterFacade userConverterFacade, ErrorMessageUtil msgUtil) {
+    public AuthenticationController(AuthenticationManager authenticationManager, UserAccountUpdateEmailer userAccountUpdateEmailer,
+            BCryptPasswordEncoder bCryptPasswordEncoder, UserManager userManager, JWTUserConverterFacade userConverterFacade,
+            ErrorMessageUtil msgUtil, CognitoAuthenticationManager cognitoAuthenticationManager, CognitoPasswordManager cognitoPasswordManager,
+            FF4j ff4j) {
         this.authenticationManager = authenticationManager;
         this.userAccountUpdateEmailer = userAccountUpdateEmailer;
         this.bCryptPasswordEncoder = bCryptPasswordEncoder;
         this.userManager = userManager;
         this.userConverterFacade = userConverterFacade;
         this.msgUtil = msgUtil;
+
+        this.cognitoAuthenticationManager = cognitoAuthenticationManager;
+        this.cognitoPasswordManager = cognitoPasswordManager;
+        this.ff4j = ff4j;
     }
 
+
+    @Operation(summary = "Log in.",
+            description = "Call this method to authenticate a user. The value returned is that user's "
+                    + "token which must be passed on all subsequent requests in the Authorization header. "
+                    + "Specifically, the Authorization header must have a value of 'Bearer token-that-gets-returned'.",
+            security = {
+                    @SecurityRequirement(name = SwaggerSecurityRequirement.API_KEY)
+            }
+        )
+    @ApiResponse(responseCode = "470", description = "The user is required to respond to the described challenge.")
+    @RequestMapping(value = "", method = RequestMethod.POST,
+            consumes = MediaType.APPLICATION_JSON_VALUE, produces = "application/json; charset=utf-8")
+    public CognitoAuthenticationResponse authenticate(@RequestBody LoginCredentials credentials) throws CognitoAuthenticationChallengeException, CognitoPasswordResetRequiredException {
+
+        if (!ff4j.check(FeatureList.SSO)) {
+            throw new NotImplementedException("This method has not been implemented");
+        }
+
+        CognitoAuthenticationResponse response = cognitoAuthenticationManager.authenticate(credentials);
+        if (response == null) {
+            throw new ChplAccountStatusException(msgUtil.getMessage("auth.loginNotAllowed"));
+        }
+        return response;
+    }
+
+    @Operation(summary = "Log user out.",
+            description = "Invalidates all of the tokens associated with the user.",
+            security = {
+                    @SecurityRequirement(name = SwaggerSecurityRequirement.API_KEY)
+            }
+        )
+    @RequestMapping(value = "/logout", method = RequestMethod.POST,
+            consumes = MediaType.APPLICATION_JSON_VALUE, produces = "application/json; charset=utf-8")
+    public void logout(@RequestBody CognitoLogoutRequest request) {
+        cognitoAuthenticationManager.invalidateTokensForUser(request.getEmail());
+    }
+
+
+    @Operation(summary = "Set user's password in response to NEW_PASSWORD_REQUIRED challenge.",
+            description = "Set user's password in response to NEW_PASSWORD_REQUIRED challenge.",
+            security = {
+                    @SecurityRequirement(name = SwaggerSecurityRequirement.API_KEY)
+            }
+        )
+    @RequestMapping(value = "/challenge/new-password-required", method = RequestMethod.POST,
+            consumes = MediaType.APPLICATION_JSON_VALUE, produces = "application/json; charset=utf-8")
+    public CognitoAuthenticationResponse newPasswordRequiredChallenge(@RequestBody CognitoNewPasswordRequiredRequest request) throws ValidationException {
+
+        if (!ff4j.check(FeatureList.SSO)) {
+            throw new NotImplementedException("This method has not been implemented");
+        }
+
+        CognitoAuthenticationResponse response = cognitoAuthenticationManager.newPassworRequiredChallenge(request);
+        if (response == null) {
+            throw new ChplAccountStatusException(msgUtil.getMessage("auth.loginNotAllowed"));
+        }
+        return response;
+    }
+
+    @Operation(summary = "Start forgot password workflow",
+            description = "Send a user an email with a link to reset their password.",
+            security = {
+                    @SecurityRequirement(name = SwaggerSecurityRequirement.API_KEY)
+            }
+        )
+    @RequestMapping(value = "/forgot-password/send-email", method = RequestMethod.POST,
+            consumes = MediaType.APPLICATION_JSON_VALUE, produces = "application/json; charset=utf-8")
+    public void requestEmailForForgottenPassword(@RequestBody CognitoForgotPasswordRequest request) throws EmailNotSentException {
+
+        if (!ff4j.check(FeatureList.SSO)) {
+            throw new NotImplementedException("This method has not been implemented");
+        }
+
+        cognitoPasswordManager.sendForgotPasswordEmail(request.getUserName());
+    }
+
+    @Operation(summary = "Complete forgot password workflow",
+            description = "Set user's password after requesting forgot password.",
+            security = {
+                    @SecurityRequirement(name = SwaggerSecurityRequirement.API_KEY)
+            }
+        )
+    @RequestMapping(value = "/forgot-password/set-password", method = RequestMethod.POST,
+            consumes = MediaType.APPLICATION_JSON_VALUE, produces = "application/json; charset=utf-8")
+    public void setForgottenPassword(@RequestBody CognitoSetForgottenPasswordRequest request) throws EmailNotSentException, ValidationException {
+        if (!ff4j.check(FeatureList.SSO)) {
+            throw new NotImplementedException("This method has not been implemented");
+        }
+
+        cognitoPasswordManager.setForgottenPassword(request.getForgotPasswordToken(), request.getPassword());
+    }
+
+    @Operation(summary = "Update the password for the currently logged in user.",
+            description = "Update the password for the currently logged in user.",
+            security = {
+                    @SecurityRequirement(name = SwaggerSecurityRequirement.API_KEY)
+            }
+        )
+    @RequestMapping(value = "/password", method = RequestMethod.POST,
+            consumes = MediaType.APPLICATION_JSON_VALUE, produces = "application/json; charset=utf-8")
+    public void setPassword(@RequestBody CognitoUpdatePasswordRequest request) throws EmailNotSentException, ValidationException, UserRetrievalException {
+        cognitoPasswordManager.setPassword(request.getPassword(), request.getConfirmPassword());
+    }
+
+    @RequestMapping(value = "/refresh-token", method = RequestMethod.POST,
+            produces = "application/json; charset=utf-8")
+    public CognitoAuthenticationResponse refreshToken(@RequestBody CognitoRefreshTokenRequest request) {
+        return cognitoAuthenticationManager.refreshAuthenticationTokens(request.getRefreshToken(), UUID.fromString(request.getCognitoId()));
+    }
+
+
+    ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
     @Deprecated
-    @DeprecatedApi(friendlyUrl = "/auth/athenticate",
+    @DeprecatedApi(friendlyUrl = "/auth/authenticate",
             httpMethod = "POST",
             removalDate = "2024-11-01",
             message = "This endpoint is deprecated and will be removed in a future release. No replacement is currently available.")
